feat: Implement MainPostList component and integrate with Redux for post management #1
