+++ src/components/Rectangle.tsx
@@ -1,6 +1,4 @@
-import { type ReactNode, useEffect, useState } from 'react';
-
-import { useTheme } from '@/hooks/useTheme';
+import { type ReactNode } from 'react';
 
 interface IRectangleProps {
   children: ReactNode;
@@ -8,22 +6,7 @@ interface IRectangleProps {
 }
 
 function Rectangle({ children, className }: IRectangleProps) {
-  const { resolvedTheme } = useTheme();
-
-  const [iconColor, setIconColor] = useState('');
-
-  useEffect(() => {
-    if (resolvedTheme === 'black') setIconColor('text-[#f9f9f9]');
-    else setIconColor('text-[#262626]');
-  }, [resolvedTheme]);
-
-  return (
-    <div
-      className={`${iconColor} ${resolvedTheme === 'black' ? 'bg-[#121212]' : 'bg-[#fafafa]'} ${className}`}
-    >
-      {children}
-    </div>
-  );
+  return <div className={`${className} bg-base-300`}>{children}</div>;
 }
 
 export default Rectangle;
