+++ src/store/index.ts
@@ -2,12 +2,14 @@ import { configureStore } from '@reduxjs/toolkit';
 
 import authReducer from './slices/authSlice';
 import counterReducer from './slices/counterSlice';
+import postReducer from './slices/postSlice';
 import socketReducer from './slices/socketSlice';
 import toastReducer from './slices/toastSlice';
 import userReducer from './slices/userSlice';
 
 export const store = configureStore({
   reducer: {
+    post: postReducer,
     counter: counterReducer,
     user: userReducer,
     auth: authReducer,
